修改类型

+++ launcher-domain/src/main/java/com/pactera/domain/LaunThemeClassificationV2.java
@@ -60,6 +60,6 @@ public class LaunThemeClassificationV2 {
     int disable;
 
     private @ApiModelProperty("格式")
-    int formatId;
+    long formatId;
 
 }
